Create and destroy swapchain system
I am still getting used to using flecs.  I was using observers for
running clean up and that is a really bad idea.  Instead I created a new
phase for the pipeline called "OnStop".  This is not built in by default
into flecs like "OnStart".  However, the concept is the same.  Instead
of using ecs.Quit when the SDL window receives a quit event, I activate
the "OnStop" phase.  This will clean up everything and then run the
ecs.quit function once Vulkan has been fully cleaned up.  We'll see how
this scales, but it works for now.