Added camera movement and rotation
This has various new systems for enabling camera movement and rotation.
This includes the ability to read keyboard and mouse events from SDL.
Then there are systems set up along with new components such as
Position, Orientation, Velocity, etc.  These components are intended to
be used well beyond just the camera system.  For instance, the Velocity
and Angular Velocity have nothing to do with the camera system
themselves.  They can simply be used to transition the current velocity
or angular velocity to a position or orientation component respectively;
this can be a camera or any other entity in the world.

There is quite a bit of clean up to do which will be done later.