Build an XCFramework for running ECS on apple platforms
This is an experiment to see how well the flecs system will run on Apple platforms.  The first challenge was building a set of libraries for some of the Apple platforms so we can make sure everything will work as expected.  This is leveraging multiple tools such as libtool, lipo, and xcframework.  The libs are processed in a very naive way right now.  The goal would be to use the command-line tools in the correct way so the framework functions like other frameworks.